Add support for relative gamma

+++ novideo_srgb/MainViewModel.cs
@@ -43,7 +43,7 @@ namespace novideo_srgb
                         (string)settings.Attribute("icc_path"),
                         (bool)settings.Attribute("calibrate_gamma"),
                         (int)settings.Attribute("selected_gamma"),
-                        (float)settings.Attribute("custom_gamma"),
+                        (double)settings.Attribute("custom_gamma"),
                         (bool?)settings.Attribute("ignore_trc") ?? false);
                 }
                 else
